feat: divided code into modules
